us852: use repo to generate next legacy application id.

+++ src/SFA.Apprenticeships.Domain.Interfaces/Repositories/IReferenceNumberRepository.cs
@@ -4,6 +4,6 @@
     {
         long GetNextVacancyReferenceNumber();
 
-        long GetNextLegacyApplicationId();
+        int GetNextLegacyApplicationId();
     }
 }
